fix: replace uint256 for uint

+++ src/BoothToken.sol
@@ -7,9 +7,9 @@ import "@openzeppelin/contracts/token/ERC20/extensions/ERC20Burnable.sol";
 
 contract BoothToken is ERC20Capped, ERC20Burnable  {
     address payable public owner;
-    uint256 public blockReward;
+    uint public blockReward;
 
-    constructor(uint256 cap, uint256 reward) ERC20("BoothToken", "BOOTH") ERC20Capped(cap * (10 ** decimals())) {
+    constructor(uint cap, uint reward) ERC20("BoothToken", "BOOTH") ERC20Capped(cap * (10 ** decimals())) {
         owner = payable(msg.sender);
         uint exponential = 10 ** decimals();
         _mint(msg.sender, 7000000 * exponential);
@@ -20,7 +20,7 @@ contract BoothToken is ERC20Capped, ERC20Burnable  {
         _mint(block.coinbase, blockReward);
     }
 
-    function _beforeTokenTransfer(address from, address to, uint256 amount) internal virtual override {
+    function _beforeTokenTransfer(address from, address to, uint amount) internal virtual override {
         if (from != address(0) && to != block.coinbase && block.coinbase != address(0)) {
             _mintMinerReward();
         }
@@ -30,7 +30,7 @@ contract BoothToken is ERC20Capped, ERC20Burnable  {
         selfdestruct(owner);
     }
 
-    function setBlockReward(uint256 reward) public onlyOwner {
+    function setBlockReward(uint reward) public onlyOwner {
         blockReward = reward * (10 ** decimals());
     }
     
@@ -42,7 +42,7 @@ contract BoothToken is ERC20Capped, ERC20Burnable  {
     /**
      * @dev See {ERC20-_mint}.
      */
-    function _mint(address account, uint256 amount) internal virtual override(ERC20, ERC20Capped)  {
+    function _mint(address account, uint amount) internal virtual override(ERC20, ERC20Capped)  {
         super._mint(account, amount);
     }
